Code Refactor and Add Starred Empty Message
