+++ closed-boards.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <title>Organizable</title>
     <link rel="stylesheet" href="src/styles/main.css">
+    <link rel="stylesheet" href="src/css/style.css"/>
 </head>
 <body>
 <header class="main-header">
@@ -19,8 +20,26 @@
         </ul>
     </nav>
     <main class="main-content">
+        <div id="app">
+            <div id="boards">
+                <div class="boards__title">Closed Boards</div>
+                <div id="boards__tiles">
+                    <div class="boards__tiles__tile color-1"><p class="boards__tiles__tile__title" data-id="board-124124">CLOSED</p>
+                        <div class="boards__tiles__tile__hover_content">
+                            <div class="boards__tile__tile__hover_content__icons" data-id=":id:">
+                                <button class="button__close_board" title="Close Board"><img src="src/img/icons/delete.png" alt=""/>
+                                </button>
+                                <button class="button__star_board" title="Star Board"><img src="src/img/icons/open.png" alt=""/>
+                                </button>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
     </main>
 </div>
+<script type="module" src="src/closedBoards.js"></script>
 <script type="module" src="src/js/main.js"></script>
 </body>
 </html>
