+++ my-boards.html
@@ -49,7 +49,8 @@
             </div>
             <div id="header">
                 <div class="header__title">Your Starred Boards</div>
-                <div id="header__starred_boards"></div>
+                <div id="header__starred_boards">
+                </div>
             </div>
             <div id="boards">
                 <div class="boards__title">Your Boards</div>
@@ -58,8 +59,7 @@
                 </div>
             </div>
         </div>
-
-        <script src="src/app.js"></script>
+        <script type="module" src="src/app.js"></script>
     </main>
 
 </div>
